backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
anthropic>=0.40.0
supabase>=2.10.0
python-dotenv>=1.0.0
boto3>=1.35.0
google-auth>=2.30.0
google-api-python-client>=2.130.0
